RideHailIterationHistory shouldn't be an Actor
