+++ src/main/scala/beam/agentsim/agents/ridehail/RideHailIterationHistory.scala
@@ -1,0 +1,41 @@
+package beam.agentsim.agents.ridehail
+
+import beam.sim.BeamServices
+import com.conveyal.r5.transit.TransportNetwork
+import javax.inject.Inject
+import org.matsim.core.api.experimental.events.EventsManager
+
+import scala.collection.mutable.ArrayBuffer
+
+class RideHailIterationHistory @Inject()(
+  eventsManager: EventsManager,
+  beamServices: BeamServices,
+  transportNetwork: TransportNetwork
+) {
+
+  //val rideHailIterationHistory=scala.collection.mutable.ListBuffer( Map[String, ArrayBuffer[Option[RideHailStatsEntry]]])
+  // TODO: put in RideHailStats class!
+  // create methods in that class, which I need for my programming
+
+  val rideHailIterationStatsHistory: ArrayBuffer[TNCIterationStats] = ArrayBuffer()
+
+  def updateRideHailStats(stats: TNCIterationStats) = {
+    rideHailIterationStatsHistory += stats
+
+    // trimm array buffer as we just need 2 elements
+    if (rideHailIterationStatsHistory.size > 2) {
+      rideHailIterationStatsHistory.remove(0)
+    }
+  }
+
+  def oszilationAdjustedTNCIterationStats(): Option[TNCIterationStats] = {
+    if (rideHailIterationStatsHistory.size >= 2) {
+      val lastElement = rideHailIterationStatsHistory.last
+      val secondLastElement = rideHailIterationStatsHistory(rideHailIterationStatsHistory.size - 2)
+      Some(TNCIterationStats.merge(lastElement, secondLastElement))
+    } else {
+      rideHailIterationStatsHistory.lastOption
+    }
+  }
+}
+
