+++ src/main/scala/beam/sim/BeamMobsim.scala
@@ -13,7 +13,7 @@ import akka.util.Timeout
 import beam.agentsim.agents.BeamAgent.Finish
 import beam.agentsim.agents.modalbehaviors.DrivesVehicle.BeamVehicleStateUpdate
 import beam.agentsim.agents.ridehail.RideHailManager.{BufferedRideHailRequestsTimeout, NotifyIterationEnds, RideHailAllocationManagerTimeout}
-import beam.agentsim.agents.ridehail.{RideHailAgent, RideHailManager, RideHailSurgePricingManager}
+import beam.agentsim.agents.ridehail.{RideHailAgent, RideHailIterationHistory, RideHailManager, RideHailSurgePricingManager}
 import beam.agentsim.agents.vehicles.EnergyEconomyAttributes.Powertrain
 import beam.agentsim.agents.vehicles._
 import beam.agentsim.agents.{BeamAgent, InitializeTrigger, Population}
@@ -53,7 +53,8 @@ class BeamMobsim @Inject()(
   val scenario: Scenario,
   val eventsManager: EventsManager,
   val actorSystem: ActorSystem,
-  val rideHailSurgePricingManager: RideHailSurgePricingManager
+  val rideHailSurgePricingManager: RideHailSurgePricingManager,
+  val rideHailIterationHistoryActor: RideHailIterationHistory
 ) extends Mobsim
     with LazyLogging
     with MetricsSupport {
@@ -151,7 +152,8 @@ class BeamMobsim @Inject()(
             beamServices.beamRouter,
             parkingManager,
             envelopeInUTM,
-            rideHailSurgePricingManager
+            rideHailSurgePricingManager,
+            rideHailIterationHistoryActor.oszilationAdjustedTNCIterationStats()
           ),
           "RideHailManager"
         )
