+++ src/main/scala/beam/sim/BeamSim.scala
@@ -9,7 +9,7 @@ import akka.actor.{ActorRef, ActorSystem, Identify}
 import akka.pattern.ask
 import akka.util.Timeout
 import beam.agentsim.agents.modalbehaviors.ModeChoiceCalculator
-import beam.agentsim.agents.ridehail.{RideHailIterationHistoryActor, TNCIterationsStatsCollector}
+import beam.agentsim.agents.ridehail.{RideHailIterationHistory, TNCIterationsStatsCollector}
 import beam.analysis.IterationStatsProvider
 import beam.analysis.plots.modality.ModalityStyleStats
 import beam.analysis.plots.GraphsStatsAgentSimEventsListener
@@ -58,10 +58,8 @@ class BeamSim @Inject()(
   private var createGraphsFromEvents: GraphsStatsAgentSimEventsListener = _
   private var modalityStyleStats: ModalityStyleStats = _
   private var expectedDisutilityHeatMapDataCollector: ExpectedMaxUtilityHeatMap = _
-  private var rideHailIterationHistoryActor: ActorRef = _
 
   private var tncIterationsStatsCollector: TNCIterationsStatsCollector = _
-  val rideHailIterationHistoryActorName = "rideHailIterationHistoryActor"
   val iterationStatsProviders: ListBuffer[IterationStatsProvider] = new ListBuffer()
   val iterationSummaryStats: ListBuffer[Map[java.lang.String, java.lang.Double]] = ListBuffer()
   var metricsPrinter: ActorRef = actorSystem.actorOf(MetricsPrinter.props())
@@ -145,14 +143,10 @@ class BeamSim @Inject()(
       beamServices.beamConfig.beam.outputs.writeEventsInterval
     )
 
-    rideHailIterationHistoryActor = actorSystem.actorOf(
-      RideHailIterationHistoryActor.props(eventsManager, beamServices, transportNetwork),
-      rideHailIterationHistoryActorName
-    )
     tncIterationsStatsCollector = new TNCIterationsStatsCollector(
       eventsManager,
       beamServices,
-      rideHailIterationHistoryActor,
+      event.getServices.getInjector.getInstance(classOf[RideHailIterationHistory]),
       transportNetwork
     )
